+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 - [Feat]: allow jvmTarget, kotlin languageVersion, and kotlin apiVersion to be specified individually in version catalog.
 - [Feat]: support Kotlin 2.x
 - [Feat]: support new compose compiler plugin for Kotlin 2.x
+- [Feat]: org.junit.platform:junit-platform-launcher added as testRuntimeOnly dependency for junit5()
+- [Feat]: kgpJunit5 version renamed to kgpJunitBom
 - [Chore]: remove kgpAndroidxComposeCompiler
 - [Chore]: update dokka to 2.0.0
 - [Chore]: update Gradle to 8.13